Handle nil tree when operation limit is hit
I noticed when setting an operation limit that go-tree-sitter would segfault.
This is because tree-sitter returns NULL when the operation limit is hit (this
also applies to the cancellation flag, but this library doesn't support that).

https://github.com/tree-sitter/tree-sitter/blob/7890a29db0b186b7b21a0a95d99fa6c562b8316b/lib/include/tree_sitter/api.h#L209-L246

I think a better solution would be to have Parse and ParseInput return (*Tree, error)
but that is a breaking API change, so I'm submitting this change for discussion.